# api/openapi.yaml
openapi: 3.0.3
info:
  title: ME CLI Sunset API
  description: |
    REST API for managing XL Axiata mobile accounts, data packages, family plans,
    and circle groups. Converted from the original Python CLI application to a
    professional Go API service with SQLite credential storage.

    ## Repository
    - **Source**: [github.com/kertasbaru/me-cli-sunset](https://github.com/kertasbaru/me-cli-sunset)

    ## Authentication Flow
    1. Request OTP via `POST /api/v1/auth/otp/request`
    2. Submit OTP via `POST /api/v1/auth/otp/submit`
    3. Use the active session for subsequent API calls

    ## Storage
    All credentials and session data are stored in a local SQLite database.
  version: 1.0.0
  contact:
    name: ME CLI Sunset
    url: https://github.com/kertasbaru/me-cli-sunset
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/packages/balance:
    get:
      tags: [Packages]
      summary: Get balance
      description: Fetches the current balance and credit information.
      operationId: getBalance
      responses:
        "200":
          description: Balance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/packages/family:
    post:
      tags: [Packages]
      summary: Get package family
      description: Fetches package options for a specific family code.
      operationId: getFamily
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyRequest"
      responses:
        "200":
          description: Family data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Family not found
  /api/v1/packages/families:
    get:
      tags: [Packages]
      summary: Get package families
      description: Fetches all package families for a given category.
      operationId: getFamilies
      parameters:
        - name: category_code
          in: query
          required: true
          schema:
            type: string
          description: Package category code
      responses:
        "200":
          description: Families data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/packages/detail:
    post:
      tags: [Packages]
      summary: Get package detail
      description: Fetches detailed information for a specific package option.
      operationId: getPackageDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageRequest"
      responses:
        "200":
          description: Package detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/packages/addons:
    get:
      tags: [Packages]
      summary: Get package addons
      description: Fetches addon packages available for a package option.
      operationId: getAddons
      parameters:
        - name: option_code
          in: query
          required: true
          schema:
            type: string
          description: Package option code
      responses:
        "200":
          description: Addons data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/packages/tiering:
    get:
      tags: [Packages]
      summary: Get tiering info
      description: Fetches the loyalty tiering information.
      operationId: getTieringInfo
      responses:
        "200":
          description: Tiering data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/packages/quotas:
    get:
      tags: [Packages]
      summary: Get quota details
      description: Fetches current quota details for sentry monitoring.
      operationId: getQuotaDetails
      responses:
        "200":
          description: Quota data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/packages/payment-methods:
    post:
      tags: [Packages]
      summary: Get payment methods
      description: Fetches available payment methods for a purchase.
      operationId: getPaymentMethods
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token_confirmation, payment_target]
              properties:
                token_confirmation:
                  type: string
                payment_target:
                  type: string
      responses:
        "200":
          description: Payment methods
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/packages/unsubscribe:
    post:
      tags: [Packages]
      summary: Unsubscribe from package
      description: Cancels an active package subscription.
      operationId: unsubscribe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quota_code, product_domain, product_subscription_type]
              properties:
                quota_code:
                  type: string
                product_domain:
                  type: string
                product_subscription_type:
                  type: string
      responses:
        "200":
          description: Unsubscribe result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  schemas:
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          description: Response data (varies by endpoint)
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        error:
          type: string
          description: Error message
...
    FamilyRequest:
      type: object
      required: [family_code]
      properties:
        family_code:
          type: string
          description: Package family code
        is_enterprise:
          type: boolean
          description: Filter by enterprise flag
        migration_type:
          type: string
          description: Migration type filter
          enum: [NONE, PRE_TO_PRIOH, PRIOH_TO_PRIO, PRIO_TO_PRIOH]
    PackageRequest:
      type: object
      required: [package_option_code]
      properties:
        package_option_code:
          type: string
          description: Package option code
        package_family_code:
          type: string
          description: Package family code
        package_variant_code:
          type: string
          description: Package variant code
...
  responses:
...
    Unauthorized:
      description: No active session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...